updated flow

+++ src/service/ApplicantService.java
@@ -141,9 +141,17 @@ public class ApplicantService {
     }
 
     public void deleteApplicantProfile() {
-        System.out.println("Deleting Applicant Profile...");
-        deleteProfileHelper();
-        CommonService.getInstance().accessLandingPage();
+        String input = Utility.inputOutput("Are you sure you want to delete your profile? (Y/N)");
+        if (input.equalsIgnoreCase("Y")) {
+            System.out.println("Deleting Applicant Profile...");
+            deleteProfileHelper();
+            CommonService.getInstance().accessLandingPage();
+        } else if (input.equalsIgnoreCase("N")) {
+            System.out.println("Profile deletion cancelled.");
+            viewApplicantProfilePage();
+        } else {
+            System.out.println("Invalid input. Profile deletion cancelled.");
+        }
     }
 
     public void updateApplicantProfile() {
